refactor(inference): split provider/ops.rs into ops/ directory
Splits the 1290-line ops.rs into four focused sub-modules under ops/:
- sanitize.rs  — secret scrubbing and error formatting helpers
- http_error.rs — HTTP error classification, Sentry routing, api_error
- models.rs    — model listing, OpenAI-compat /models parsing
- provider_factory.rs — provider construction (create_*, ProviderRuntimeOptions)

External API preserved exactly via pub use re-exports in ops/mod.rs.
Test companion ops_tests.rs stays in place via #[path] attribute.
