+++ src/openhuman/inference/provider/ops/sanitize.rs
@@ -1,0 +1,88 @@
+use crate::openhuman::util;
+
+pub const MAX_API_ERROR_CHARS: usize = 200;
+const TRANSPORT_ERROR_MAX_CHARS: usize = 1200;
+
+fn is_secret_char(c: char) -> bool {
+    c.is_ascii_alphanumeric() || matches!(c, '-' | '_' | '.' | ':')
+}
+
+fn token_end(input: &str, from: usize) -> usize {
+    let mut end = from;
+    for (i, c) in input[from..].char_indices() {
+        if is_secret_char(c) {
+            end = from + i + c.len_utf8();
+        } else {
+            break;
+        }
+    }
+    end
+}
+
+/// Scrub known secret-like token prefixes from provider error strings.
+pub fn scrub_secret_patterns(input: &str) -> String {
+    const PREFIXES: [&str; 7] = [
+        "sk-",
+        "xoxb-",
+        "xoxp-",
+        "ghp_",
+        "gho_",
+        "ghu_",
+        "github_pat_",
+    ];
+
+    let mut scrubbed = input.to_string();
+
+    for prefix in PREFIXES {
+        let mut search_from = 0;
+        loop {
+            let Some(rel) = scrubbed[search_from..].find(prefix) else {
+                break;
+            };
+
+            let start = search_from + rel;
+            let content_start = start + prefix.len();
+            let end = token_end(&scrubbed, content_start);
+
+            if end == content_start {
+                search_from = content_start;
+                continue;
+            }
+
+            scrubbed.replace_range(start..end, "[REDACTED]");
+            search_from = start + "[REDACTED]".len();
+        }
+    }
+
+    scrubbed
+}
+
+/// Sanitize API error text by scrubbing secrets and truncating length.
+pub fn sanitize_api_error(input: &str) -> String {
+    let scrubbed = scrub_secret_patterns(input);
+    util::truncate_with_ellipsis(&scrubbed, MAX_API_ERROR_CHARS)
+}
+
+/// Full `source()` chain for connection / TLS failures (scrubbed, longer than API body snippets).
+pub fn format_error_chain(err: &dyn std::error::Error) -> String {
+    let mut parts: Vec<String> = vec![err.to_string()];
+    let mut src = std::error::Error::source(err);
+    while let Some(e) = src {
+        parts.push(e.to_string());
+        src = std::error::Error::source(e);
+    }
+    let joined = parts.join(" | ");
+    let scrubbed = scrub_secret_patterns(&joined);
+    util::truncate_with_suffix(&scrubbed, TRANSPORT_ERROR_MAX_CHARS, "…")
+}
+
+/// Cause chain from [`anyhow::Error`] (e.g. responses fallback), scrubbed and length-limited.
+pub fn format_anyhow_chain(err: &anyhow::Error) -> String {
+    let joined = err
+        .chain()
+        .map(|e| e.to_string())
+        .collect::<Vec<_>>()
+        .join(" | ");
+    let scrubbed = scrub_secret_patterns(&joined);
+    util::truncate_with_suffix(&scrubbed, TRANSPORT_ERROR_MAX_CHARS, "…")
+}
